requirements.txt
pytest
allure-pytest
selene==2.0.0rc9
requests
python-dotenv
appium-python-client